feat: NFC card school_id, license-server auth, settings schema validation, docs audit fixes
NFC V2 cards (student_cards, staff_cards, parent_cards, employee_cards):
- Added school_id column (nullable FK) with migration d9e8f7a6b5c4
- Service assign_* functions now populate school_id from parent entity
- Response schemas include school_id field

License server:
- POST /verify, POST /activate, GET /school/{school_id} now require JWT auth

Settings PUT:
- Replaced unrestricted dict with SchoolSettingsUpdate Pydantic model
- Unknown keys rejected via extra='forbid'

IGA endpoints:
- Added INFRASTRUCTURE_VIEW permission, replacing AUDIT_VIEW

QR encryption:
- AES-256-GCM with HKDF-derived key, backward compat for base64 tokens
- New tokens prefixed A1|

Code health:
- Duplicate _is_token_blacklisted removed (consolidated in auth_deps)
- Last 2 Float money columns converted to DECIMAL(15,2)
- 173/173 tests pass (up from 159)

+++ backend/app/models/inventory_asset.py
@@ -1,6 +1,6 @@
 import uuid
 from datetime import datetime, timezone
-from sqlalchemy import Column, String, Float, DateTime, ForeignKey
+from sqlalchemy import Column, String, DECIMAL, DateTime, ForeignKey
 from app.database import Base
 
 
@@ -11,7 +11,7 @@ class InventoryAsset(Base):
     school_id = Column(String(36), ForeignKey("schools.id"), nullable=False)
     name = Column(String(255), nullable=False)
     category = Column(String(100), nullable=True)
-    value = Column(Float, default=0)
+    value = Column(DECIMAL(15, 2), default=0.00)
     location = Column(String(255), nullable=True)
     status = Column(String(20), default="active")
     created_at = Column(DateTime, default=lambda: datetime.now(timezone.utc))
